atomisp2:css2400 package  irci_master_20140712_0614
BZ: 210970

Merge CSS2.1 for ISP2400 for 'irci_master_20140712_0614' firmware tag.

+++ drivers/media/pci/atomisp2/css2400b0_v21/ia_css_device_access.c
@@ -92,11 +92,11 @@ ia_css_device_store_uint64(const hrt_address addr, const uint64_t data)
 void
 ia_css_device_load(const hrt_address addr, void *data, const size_t size)
 {
-	my_env.load(addr, data, size);
+	my_env.load(addr, data, (uint32_t)size);
 }
 
 void
 ia_css_device_store(const hrt_address addr, const void *data, const size_t size)
 {
-	my_env.store(addr, data, size);
+	my_env.store(addr, data, (uint32_t)size);
 }
